organizing and dockerfile fixes
