+++ services/grpc/config.toml
@@ -1,0 +1,3 @@
+PORT = "50051"
+FILE = "data.db"
+DIALECT = "sqlite3"
