+++ services/http/server.go
@@ -3,11 +3,11 @@ package main
 import (
 	"log"
 
+	"github.com/agrimrules/cncf/services/config"
 	"github.com/jinzhu/gorm"
 	_ "github.com/jinzhu/gorm/dialects/sqlite"
 	"github.com/kataras/iris"
 	"github.com/kataras/iris/context"
-	"github.com/spf13/viper"
 )
 
 // Post is a basic struct indicating a user and message
@@ -17,31 +17,9 @@ type Post struct {
 	Message string `json:"message"`
 }
 
-// Configuration is a struct that stores basic config objects
-type Configuration struct {
-	PORT    string
-	DIALECT string
-	URL     string
-}
-
-func initConfig() (Configuration, error) {
-	viper.SetConfigName("config")
-	viper.AddConfigPath(".")
-	err := viper.ReadInConfig()
-	if err != nil {
-		return Configuration{}, err
-	}
-	viper.SetDefault("PORT", "8080")
-	viper.SetDefault("DIALECT", "sqlite3")
-	viper.SetDefault("URL", "data.db")
-	var config Configuration
-	err = viper.Unmarshal(&config)
-	return config, err
-}
-
 func main() {
 	app := iris.Default()
-	conf, err := initConfig()
+	conf, err := config.InitConfig()
 	if err != nil {
 		log.Fatal(err)
 		panic("Unable to load config")
